refactor extractFeature.py to get more flexibility for adding multi-window feature extraction
