+++ src/extractFeature.py
@@ -1,151 +1,21 @@
 import readData
+
+import numpy as np
+import pandas as pd
+
 augmentationName = readData.sensor_name[:]
 originalName = readData.sensor_name[:]
 augmentationName.append('acc')
 augmentationName.append('gxyz')
 
-import numpy as np
-import pandas as pd
-class Normalizer:
-    def __init__(self, normalizer_type,data):
-        print "********** init normalizer **********"
-        normalizer_map = {
-                "zscore1free":self.normalizeDataMeanStd1free,
-                "zscore0free":self.normalizeDataMeanStd0free,
-                "minmax":self.normalizeDataMinMax,
-                "robust":self.normalizeDataRobust
-        }
-        from sklearn import preprocessing
-        self.normalizer = normalizer_map[normalizer_type]
-        self.standrad_scaler = preprocessing.StandardScaler().fit(data)
-        self.minmax_scaler = preprocessing.MinMaxScaler().fit(data)
-        self.robust_scaler = preprocessing.RobustScaler().fit(data)
-        
-    def normalizeDataMeanStd1free(self,ori_data):
-        print "********** MeanStd1free normalize data ***********"
-        if type(ori_data) is list:
-            data = ori_data[:]
-            for i in range(len(ori_data)):
-                data[i] = (ori_data[i] - ori_data[i].mean())/ori_data[i].std()
-        else:
-            data = ori_data.copy()
-            data = (ori_data - ori_data.mean()) / ori_data.std()
-        return data
-        
-    def normalizeDataMeanStd0free(self,ori_data):
-        print "********** MeanStd0free normalize data ***********"
-        if type(ori_data) is list:
-            data = ori_data[:]
-            for i in range(len(ori_data)):
-                data[i] = pd.DataFrame(self.standrad_scaler.transform(data[i]),
-                                    index=data[i].index,columns=data[i].columns)
-        else:
-            data = ori_data.copy()
-            data = pd.DataFrame(self.standrad_scaler.transform(data),
-                                 index=data.index,columns=data.columns)
-        return data
-        
-    def normalizeDataMinMax(self,ori_data):
-        print "********** Minmax normalize data ***********"
-        if type(ori_data) is list:
-            data = ori_data[:]
-            for i in range(len(ori_data)):
-                data[i] = pd.DataFrame(self.minmax_scaler.transform(data[i]),
-                                index=data[i].index,columns=data[i].columns)
-        else:
-            data = ori_data.copy()
-            data = pd.DataFrame(self.minmax_scaler.transform(data),
-                                index=data.index,columns=data.columns)
-        return data
-    
-    def normalizeDataRobust(self,ori_data):
-        print "********** Robust normalize data ***********"
-        if type(ori_data) is list:
-            data = ori_data[:]
-            for i in range(len(ori_data)):
-                data[i] = pd.DataFrame(self.robust_scaler.transform(data[i]),
-                                index=data[i].index,columns=data[i].columns)
-        else:
-            data = ori_data.copy()
-            data = pd.DataFrame(self.robust_scaler.transform(data),
-                                index=data.index,columns=data.columns)
-        return data
-        
-# end of class normalizer define
-        
-class PCAor:
-    def __init__(self,pca_type,data,n_components):
-        print "******** init PCAer **********"
-        pca_map = {
-            "already":self.pcawithalreadydata,
-            "normal":self.normalPCA,
-            "kernal":self.kernalPCA,
-        }
-        self.components_index = []
-        self.eig_vecs = []
-        self.pcaor = pca_map[pca_type]
-        
-        # normal pca init and fit
-        from sklearn.decomposition import PCA
-        self.normal_pca = PCA(n_components=n_components,whiten=False)
-        self.normal_pca.fit(data)
-        # already manually standarize data then pca init and fit
-        self._pcawithalreadydatatofit(data,n_components)
-        # kernal pca init and fit        
-        from sklearn.decomposition import KernelPCA
-        self.kernal_pca = KernelPCA(kernel="cosine", gamma=10) #cosine for knn
-        self.kernal_pca.fit(data)
-    
-    # to fit the data then get the principle components
-    def _pcawithalreadydatatofit(self,ori_data,n_components):
-        print "********** Fit Already data ***********"
-        cov_mat = ori_data.T.dot(ori_data)/(ori_data.shape[0]-1)
-        eig_vals, eig_vecs = np.linalg.eig(cov_mat)
-        tot = eig_vals.sum(axis = 0)
-        var_exp = pd.DataFrame(eig_vals/tot)
-        var_exp = var_exp.sort_values(by=0, ascending=False)
-        self.components_index = var_exp.index[0:n_components]
-        self.eig_vecs = eig_vecs[self.components_index]
-    
-    def pcawithalreadydata(self,ori_data):
-        print "********** Transform Already data ***********"
-        if type(ori_data) is list:
-            data = ori_data[:]
-            for i in range(len(ori_data)):
-                data[i] = ori_data[i].dot(self.eig_vecs.T)
-        else:
-            data = ori_data.dot(self.eig_vecs.T)
-        return data
-    
-    def normalPCA(self,ori_data):
-        print "********** Normal PCA Transform ***********"  
-        if type(ori_data) is list:
-            data = ori_data[:]
-            for i in range(len(ori_data)):
-                data[i] = self.normal_pca.transform(ori_data[i])
-        else:
-            data = self.normal_pca.transform(ori_data)
-        return data
-    
-    def kernalPCA(self,ori_data):
-        print "********** Kernal PCA Transform ***********"
-        if type(ori_data) is list:
-            data = ori_data[:]
-            for i in range(len(ori_data)):
-                data[i] = self.kernal_pca.transform(ori_data[i])
-        else:
-            data = self.kernal_pca.transform(ori_data)
-        return data
-# end of class PCAor define
-       
 copydata = 0
 flag = 1
-class featureExtractor:
+class FeatureExtractor:
     def __init__(self):
-        print "******** featureExtractor init **********"
+        print "******** FeatureExtractor init **********"
         self.sensor_name = readData.sensor_name
         
-    def ExtractTraditonFeature(self,data):
+    def _ExtractTraditonFeature(self,data):
         data = pd.concat([data],ignore_index=True)
         m_absData = np.abs(data)
         step = 50
@@ -244,17 +114,8 @@ class featureExtractor:
         m_featrue.reset_index(drop=True,inplace=True)
         m_featrue = m_featrue.T
         return m_featrue
-        
-    def ExtractFeatureinShipengStyle(self,m_unnormalizedData,m_startPoints,needBanlance = True):
-        print "********** extract feature **********"        
-        UnnormalizedData = m_unnormalizedData.copy()
-        UnnormalizedData[self.sensor_name[0]] = UnnormalizedData[self.sensor_name[0]] / 2048
-        UnnormalizedData[self.sensor_name[1]] = UnnormalizedData[self.sensor_name[1]] / 2048
-        UnnormalizedData[self.sensor_name[2]] = UnnormalizedData[self.sensor_name[2]] / 2048
-        UnnormalizedData[self.sensor_name[3]] = UnnormalizedData[self.sensor_name[3]] / 1879.44
-        UnnormalizedData[self.sensor_name[4]] = UnnormalizedData[self.sensor_name[4]] / 1879.44
-        UnnormalizedData[self.sensor_name[5]] = UnnormalizedData[self.sensor_name[5]] / 1879.44
-            
+    
+    def _DealListData(self,m_data,m_startPoints,needBanlance): 
         windowWidth = 100
         if needBanlance:
             # if need banlance the data, num limitation set to 50,
@@ -264,78 +125,64 @@ class featureExtractor:
             numLimit = 999999
         if type(m_startPoints[0]) is list:
             featureOfSensor = []
-            for i in range(len(m_startPoints)-1):
+            for i in range(len(m_startPoints)):
                 featureOfSensor.append(pd.DataFrame())
-            for classIndex in range(len(m_startPoints)-1):
+            for classIndex in range(len(m_startPoints)):
                 numCount = 0
                 for startPos in m_startPoints[classIndex]:
                     if numCount < numLimit:
                         numCount = numCount + 1
-                        tempFeature = self.ExtractTraditonFeature(
-                            UnnormalizedData.loc[startPos:startPos+windowWidth-1])
-                        featureOfSensor[classIndex] = featureOfSensor[classIndex].append(tempFeature)
-                    else:
-                        break
-                featureOfSensor[classIndex].reset_index(drop=True,inplace=True)
-#               import matplotlib.pyplot as plt
-#               plt.figure()
-#               plt.plot(featureOfSensor[classIndex][0])
-        else:
-            featureOfSensor = pd.DataFrame()    
-            numCount = 0
-            for startPos in m_startPoints:
-                if numCount < numLimit:
-                    numCount = numCount + 1
-                    tempFeature = self.ExtractTraditonFeature(
-                        UnnormalizedData.loc[startPos:startPos+windowWidth-1])
-                    featureOfSensor = featureOfSensor.append(tempFeature)
-            featureOfSensor.reset_index(drop=True,inplace=True)
-        return featureOfSensor
-
-# end of class featureExtractor define
-        
-class AdvancedFeatureExtractor(featureExtractor):
-    def __init__(self):
-        self.sensor_name = readData.sensor_name
-        print "****** AdvancedFeatureExtractor init ********"
-
-    def ExtractFeatureinShipengStyle(self,m_normalizedData,m_startPoints,needBanlance = True):
-        print "********** extract advance feature **********"
-        m_data = m_normalizedData.copy()  
-        windowWidth = 100      
-        if needBanlance:
-            # if need banlance the data, num limitation set to 50,
-            # so that (number of shoot samples etc.) = (number of run samples etc.) 
-            numLimit = 50
-        else:
-            numLimit = 999999
-        if type(m_startPoints[0]) is list:
-            featureOfSensor = []
-            for i in range(len(m_startPoints)-1):
-                featureOfSensor.append(pd.DataFrame())
-            for classIndex in range(len(m_startPoints)-1):
-                numCount = 0
-                for startPos in m_startPoints[classIndex]:
-                    if numCount < numLimit:
-                        numCount = numCount + 1
-                        tempFeature = self.ExtractTraditonFeature(
+                        tempFeature = self._ExtractTraditonFeature(
                             m_data.loc[startPos:startPos+windowWidth-1])
                         featureOfSensor[classIndex] = featureOfSensor[classIndex].append(tempFeature)
                     else:
                         break
                 featureOfSensor[classIndex].reset_index(drop=True,inplace=True)
+        return featureOfSensor
+
+# end of class FeatureExtractor define
+
+class UnnormaDatafeatureExtractor(FeatureExtractor):
+    def __init__(self):
+        print "******** UnnormaDatafeatureExtractor init **********"
+        self.sensor_name = readData.sensor_name
+    def ExtractFeatureinShipengStyle(self,m_unnormalizedData,m_startPoints,needBanlance = True):
+        print "********** extract feature from unpreNorma data **********"        
+        UnnormalizedData = m_unnormalizedData.copy()
+        UnnormalizedData[self.sensor_name[0]] = UnnormalizedData[self.sensor_name[0]] / 2048
+        UnnormalizedData[self.sensor_name[1]] = UnnormalizedData[self.sensor_name[1]] / 2048
+        UnnormalizedData[self.sensor_name[2]] = UnnormalizedData[self.sensor_name[2]] / 2048
+        UnnormalizedData[self.sensor_name[3]] = UnnormalizedData[self.sensor_name[3]] / 1879.44
+        UnnormalizedData[self.sensor_name[4]] = UnnormalizedData[self.sensor_name[4]] / 1879.44
+        UnnormalizedData[self.sensor_name[5]] = UnnormalizedData[self.sensor_name[5]] / 1879.44
+        m_data = UnnormalizedData.copy()
+        
+        if type(m_startPoints[0]) is list:
+            featureOfSensor = self._DealListData(m_data,m_startPoints,needBanlance)
         else:
-            featureOfSensor = pd.DataFrame()
-            numCount = 0
-            for startPos in m_startPoints:
-                if numCount < numLimit:
-                    numCount = numCount + 1
-                    tempFeature = self.ExtractTraditonFeature(
-                        m_data.loc[startPos:startPos+windowWidth-1])
-                    featureOfSensor = featureOfSensor.append(tempFeature)
-                else:
-                    break
-            featureOfSensor.reset_index(drop=True,inplace=True)
+            featureOfSensor = self._DealListData(m_data,[m_startPoints],needBanlance)
+            featureOfSensor = featureOfSensor[0]
+        return featureOfSensor
+
+# end of class UnnormaDatafeatureExtractor define
+
+class NormaDatafeatureExtractor(FeatureExtractor):
+    def __init__(self):
+        self.sensor_name = readData.sensor_name
+        print "****** NormaDatafeatureExtractor init ********"
+
+# ExtractFeatureinShipengStyle in NormaDatafeatureExtractor
+# use normalizeData from readData()
+# instead use unnormalizeData then div 2048 & 1879
+    def ExtractFeatureinShipengStyle(self,m_normalizedData,m_startPoints,needBanlance = True):
+        print "********** extract feature from preNorma data **********"
+        m_data = m_normalizedData.copy()  
+        
+        if type(m_startPoints[0]) is list:
+            featureOfSensor = self._DealListData(m_data,m_startPoints,needBanlance)
+        else:
+            featureOfSensor = self._DealListData(m_data,[m_startPoints],needBanlance)
+            featureOfSensor = featureOfSensor[0]
         return featureOfSensor
         
-# end of class AdvancedFeatureExtractor define
+# end of class NormaDatafeatureExtractor define
